// github/types/paths/notifications/threads/{thread_id}.types.ts
// This code was automatically generated from an OpenAPI description.
// Do not edit this file. Edit the OpenAPI file instead.
// For more information, see https://github.com/pmcelhaney/counterfact/blob/main/docs/faq.md

import type { WideOperationArgument } from "../../../../counterfact-types/index.ts";
import type { OmitValueWhenNever } from "../../../../counterfact-types/index.ts";
import type { MaybePromise } from "../../../../counterfact-types/index.ts";
import type { COUNTERFACT_RESPONSE } from "../../../../counterfact-types/index.ts";
import type { Context } from "../../../../routes/_.context.ts";
import type { ResponseBuilderFactory } from "../../../../counterfact-types/index.ts";
import type { thread } from "../../../components/schemas/thread.js";
import type { not_modified } from "../../../#/components/responses/not_modified.js";
import type { requires_authentication } from "../../../#/components/responses/requires_authentication.js";
import type { forbidden } from "../../../#/components/responses/forbidden.js";

/**
 * Gets information about a notification thread.
 */
export type activityGetThread = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: activityGetThread_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: thread;
          };
        };
        examples: {
          default: unknown;
        };
      };
      304: not_modified;
      401: requires_authentication;
      403: forbidden;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * Marks a thread as "read." Marking a thread as "read" is equivalent to clicking a notification in your notification inbox on GitHub: https://github.com/notifications.
 */
export type activityMarkThreadAsRead = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: activityMarkThreadAsRead_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      205: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      304: not_modified;
      403: forbidden;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * Marks a thread as "done." Marking a thread as "done" is equivalent to marking a notification in your notification inbox on GitHub as done: https://github.com/notifications.
 */
export type activityMarkThreadAsDone = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: activityMarkThreadAsDone_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type activityGetThread_Path = { thread_id: number };

export type activityMarkThreadAsRead_Path = { thread_id: number };

export type activityMarkThreadAsDone_Path = { thread_id: number };